refactor: header, footer 구조 변경
