+++ src/components/CrewDetail/CrewDetail.style.js
@@ -1,5 +1,8 @@
 import styled from 'styled-components';
 
 export const StyledCrewDetail = styled.div`
+  width: 100%;
+  max-width: 1000px;
+  margin: 60px auto;
   padding: 20px;
 `;
