+++ src/components/layout/Footer.jsx
@@ -1,16 +1,17 @@
-import React from "react";
-import { styled } from "styled-components";
+import {
+  StyledFooter,
+  LinkContainer,
+  StyledLink } from "./Footer.style";
 
-const Footer = () => {
-  return <Footersection>Footer</Footersection>;
+function Footer() {
+  return (
+    <StyledFooter>
+      <LinkContainer>
+        <StyledLink to='/login'>로그인</StyledLink>
+        <StyledLink to='/register'>회원가입</StyledLink>
+      </LinkContainer>
+    </StyledFooter>
+  );
 };
 
 export default Footer;
-
-const Footersection = styled.footer`
-
-  background-color: wheat; /* Add your desired background color */
-  padding: 20px; /* Add any desired padding */
-  text-align: center; /* Center align the content within the footer */
-  font-size: 30px;
-`;
