+++ src/components/layout/MainLayout.jsx
@@ -17,6 +17,7 @@ const MainLayout = ({ children }) => {
 
 export default MainLayout;
 const MainLayoutSection = styled.div`
+  margin-top: 60px;
   margin-left: auto;
   margin-right: auto;
   width: 1000px;
