+++ src/pages/CrewDetailPage.jsx
@@ -1,7 +1,15 @@
+import Header from '../components/layout/Header';
 import CrewDetail from '../components/CrewDetail/CrewDetail';
+import Footer from '../components/layout/Footer';
 
 function CrewDetailPage() {
-  return <CrewDetail />;
+  return (
+    <>
+      <Header />
+      <CrewDetail />
+      <Footer />
+    </>
+  )
 };
 
 export default CrewDetailPage;
